requirements.txt
requests
wheel
mysql.connector
pandas
opencv-python
numpy